synch --live-replay feature with playitagainsam, or maybe not
Not planning a --live-replay feature for playitagainsam.js,
though it would be easy to do by remote serving the actual terminal
and adding a --server --port=localport option to playitagainsam.

Then, the js call would be something like:

    player.live-replay("./path/to/terminal-session.js", localport);

But I'm not going to do it, because that's wholly unnecessary.

I'm just doing this because it's always good to comment on feature
parity between compatible projects.

And to bump playitagainsam up in your queue :).